[issue-21645] handle corner cases - improve code to not crash on message non-existence

+++ record-and-playback/core/lib/recordandplayback/generators/events.rb
@@ -842,14 +842,14 @@ module BigBlueButton
 
           # Properly delete this message
           index_to_be_deleted = chats.index { |message| message[:id] === message_id }
-          chats.delete_at(index_to_be_deleted)
+          chats.delete_at(index_to_be_deleted) unless index_to_be_deleted.nil?
         when %w[CHAT EditPublicChatMessageRecordEvent]
           next if timestamp < start_time
           message_id = event.at_xpath('./messageId')&.content
           new_message_content = event.at_xpath('./message')&.content
-          index_to_be_deleted = chats.index { |message| message[:id] === message_id }
-          chats[index_to_be_deleted][:message] = new_message_content
-          chats[index_to_be_deleted][:lastEditedTimestamp] = timestamp
+          index_to_be_edited = chats.index { |message| message[:id] === message_id }
+          chats[index_to_be_edited][:message] = new_message_content unless index_to_be_edited.nil?
+          chats[index_to_be_edited][:lastEditedTimestamp] = timestamp unless index_to_be_edited.nil?
         when %w[PARTICIPANT RecordStatusEvent]
           record = event.at_xpath('status').content == 'true'
           next if timestamp < start_time
